move FadeIn to lib/Components
